Refactor index reducer
