+++ packages/wallet/src/redux/reducers/index.ts
@@ -1,40 +1,14 @@
-import {
-  OPENING,
-  FUNDING,
-  RUNNING,
-  CHALLENGING,
-  RESPONDING,
-  WITHDRAWING,
-  CLOSING,
-  approveConclude,
-  ApproveConclude,
-  acknowledgeConclude,
-  AcknowledgeConclude,
-  ChannelState,
-} from '../states/channels';
 import { WalletState, INITIALIZING, waitForLogin, WALLET_INITIALIZED } from '../states';
 
 import { initializingReducer } from './initializing';
-import { openingReducer } from './channels/opening';
-import { fundingReducer } from './channels/funding';
-import { runningReducer } from './channels/running';
-import { challengingReducer } from './channels/challenging';
-import { respondingReducer } from './channels/responding';
-import { withdrawingReducer } from './channels/withdrawing';
-import { closingReducer } from './channels/closing';
 import {
   WalletAction,
-  CONCLUDE_REQUESTED,
   MESSAGE_SENT,
   TRANSACTION_SENT_TO_METAMASK,
   DISPLAY_MESSAGE_SENT,
-  COMMITMENT_RECEIVED,
 } from '../actions';
-import { unreachable, ourTurn, validTransition } from '../../utils/reducer-utils';
-import { validCommitmentSignature } from '../../utils/signing-utils';
-import { showWallet } from 'magmo-wallet-client/lib/wallet-events';
-import { CommitmentType } from 'fmg-core';
-import { OutboxState } from '../states/shared';
+import { unreachable } from '../../utils/reducer-utils';
+import { OutboxState, SharedWalletState } from '../states/shared';
 import { initializedReducer } from './initialized';
 
 const initialState = waitForLogin();
@@ -53,93 +27,30 @@ export const walletReducer = (
   if (action.type === TRANSACTION_SENT_TO_METAMASK) {
     nextOutbox.transactionOutbox = undefined;
   }
-  state = {
-    ...state,
-    outboxState: outboxReducer(state.outboxState, { messageOutbox: undefined }),
-  };
-
-  if (state.stage === WALLET_INITIALIZED) {
-    const conclusionStateFromOwnRequest = receivedValidOwnConclusionRequest(
-      state.channelState,
-      action,
-    );
-    if (conclusionStateFromOwnRequest) {
-      return { ...state, channelState: conclusionStateFromOwnRequest };
-    }
-  }
-
-  const conclusionStateFromOpponentRequest = receivedValidOpponentConclusionRequest(state, action);
-  if (conclusionStateFromOpponentRequest) {
-    return { ...state, channelState: conclusionStateFromOpponentRequest };
-  }
+  state = sideEffectsReducer(state, nextOutbox);
 
   switch (state.stage) {
     case INITIALIZING:
       return initializingReducer(state, action);
     case WALLET_INITIALIZED:
       return initializedReducer(state, action);
-    // case OPENING:
-    //   return openingReducer(state.channelState, action);
-    // case FUNDING:
-    //   return fundingReducer(state, action);
-    // case RUNNING:
-    //   return runningReducer(state, action);
-    // case CHALLENGING:
-    //   return challengingReducer(state, action);
-    // case RESPONDING:
-    //   return respondingReducer(state, action);
-    // case WITHDRAWING:
-    //   return withdrawingReducer(state, action);
-    // case CLOSING:
-    //   return closingReducer(state, action);
     default:
       return unreachable(state);
   }
 };
 
-const receivedValidOwnConclusionRequest = (
-  state: InitializingChannelState | ChannelState,
-  action: WalletAction,
-): ApproveConclude | null => {
-  if (state.stage !== FUNDING && state.stage !== RUNNING) {
-    return null;
-  }
-  if (action.type !== CONCLUDE_REQUESTED || !ourTurn(state)) {
-    return null;
-  }
-  return approveConclude({ ...state, displayOutbox: showWallet() });
-};
+export function sideEffectsReducer<T extends SharedWalletState>(
+  state: T,
+  sideEffects: OutboxState | undefined,
+): T {
+  // TODO: What about unhandled actions?
 
-const receivedValidOpponentConclusionRequest = (
-  state: WalletState,
-  action: WalletAction,
-): AcknowledgeConclude | null => {
-  if (state.stage !== FUNDING && state.stage !== RUNNING) {
-    return null;
+  if (!sideEffects) {
+    return state;
   }
-  if (action.type !== COMMITMENT_RECEIVED) {
-    return null;
-  }
+  // TODO: We need to think about whether to overwrite existing outbox items.
+  const newState = { ...state, outboxState: { ...state.outboxState } };
+  Object.keys(sideEffects).map(k => (newState.outboxState[k] = sideEffects[k]));
 
-  const { commitment, signature } = action;
-
-  if (commitment.commitmentType !== CommitmentType.Conclude) {
-    return null;
-  }
-  // check signature
-  const opponentAddress = state.participants[1 - state.ourIndex];
-  if (!validCommitmentSignature(commitment, signature, opponentAddress)) {
-    return null;
-  }
-  if (!validTransition(state, commitment)) {
-    return null;
-  }
-
-  return acknowledgeConclude({
-    ...state,
-    turnNum: commitment.turnNum,
-    lastCommitment: { commitment, signature },
-    penultimateCommitment: state.lastCommitment,
-    displayOutbox: showWallet(),
-  });
-};
+  return newState;
+}
